- added new tests for EnvironmentServiceImplTest and AnimalEnvironmentServiceImplTest - added IllegalArgumentException throwing for null arguments in AnimalEnvironmentService and EnvironmentService - updated equals() for Animal and AnimalEaten entity

+++ DominatingSpeciesService/src/main/java/cz/muni/fi/pa165/dominatingspecies/service/AnimalEnvironmentServiceImpl.java
@@ -19,6 +19,9 @@ public class AnimalEnvironmentServiceImpl implements AnimalEnvironmentService {
 
     @Override
     public void create(AnimalEnvironment animalEnvironment) {
+        if (animalEnvironment == null) {
+            throw new IllegalArgumentException("AnimalEnvironment can not be created, parameter null.");
+        }
         animalEnvironmentDao.create(animalEnvironment);
     }
 
@@ -34,11 +37,17 @@ public class AnimalEnvironmentServiceImpl implements AnimalEnvironmentService {
 
     @Override
     public void remove(AnimalEnvironment animalEnvironment) {
+        if (animalEnvironment == null) {
+            throw new IllegalArgumentException("AnimalEnvironment can not be removed, parameter null.");
+        }
         animalEnvironmentDao.remove(animalEnvironment);
     }
 
     @Override
     public void update(AnimalEnvironment animalEnvironment) {
+        if (animalEnvironment == null) {
+            throw new IllegalArgumentException("AnimalEnvironment can not be removed, parameter null.");
+        }
         animalEnvironmentDao.update(animalEnvironment);
     } 
 }
